LayoutTests/imported/w3c/web-platform-tests/svg/pservers/scripted/stop-color-inheritance-currentcolor-expected.txt


PASS "currentcolor" is inherited as a keyword
